Adds better test coverage - Draws prop types from a unified source
